move/rename old script
